# public/openapi.yaml
openapi: 3.0.3
info:
  title: AgentsFlow API
  description: |
    API de automação de marketing via WhatsApp.

    ## Autenticação

    Todas as rotas protegidas exigem autenticação via:
    - **Session Cookie** (para navegador/UI)
    - **Bearer Token**: `Authorization: Bearer {AGENTSFLOW_API_KEY}`
    - **API Key Header**: `X-API-Key: {AGENTSFLOW_API_KEY}`

    ## Rotas Públicas

    As seguintes rotas NÃO exigem autenticação:
    - `/api/auth/*` - Login/logout
    - `/api/webhook` - Webhooks da Meta
    - `/api/health` - Health check
    - `/api/system` - Info do sistema
    - `/api/installer/*` - Setup inicial
  version: 1.0.0
  contact:
    name: AgentsFlow
  license:
    name: Proprietary
    url: https://agentsflow.app/license
servers:
  - url: http://localhost:3000
    description: Desenvolvimento
  - url: https://{domain}
    description: Produção
    variables:
      domain:
        default: seu-app.vercel.app
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Use AGENTSFLOW_API_KEY como token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Use AGENTSFLOW_API_KEY
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/webhook:
    get:
      tags: [System]
      summary: Validação do webhook (Meta Challenge)
      security: []
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Challenge aceito
        '403':
          description: Token inválido

    post:
      tags: [System]
      summary: Receber eventos da Meta
      security: []
      description: |
        Webhook para receber mensagens e status de entrega da Meta.
        Usa verificação HMAC-SHA256.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Evento processado